# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BikerBlog API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Author {
  id: number;
  name: string;
  avatarUrl: string;
  bio: string;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  bike?: string | null;
}

export interface AuthorSummary {
  id: number;
  name: string;
  avatarUrl: string;
  bio?: string;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  bike?: string | null;
  postCount: number;
}

export type PostCategory = typeof PostCategory[keyof typeof PostCategory];


export const PostCategory = {
  'ride-report': 'ride-report',
  'gear-review': 'gear-review',
  'tech-tips': 'tech-tips',
  news: 'news',
} as const;

export interface Post {
  id: number;
  slug: string;
  title: string;
  excerpt: string;
  /** Markdown body */
  content: string;
  coverImageUrl: string;
  category: PostCategory;
  tags: string[];
  author: Author;
  publishedAt: string;
  readingMinutes: number;
  likeCount: number;
  commentCount: number;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  bike?: string | null;
}

export type PostInputCategory = typeof PostInputCategory[keyof typeof PostInputCategory];


export const PostInputCategory = {
  'ride-report': 'ride-report',
  'gear-review': 'gear-review',
  'tech-tips': 'tech-tips',
  news: 'news',
} as const;

export interface PostInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  excerpt: string;
  /** @minLength 1 */
  content: string;
  coverImageUrl: string;
  category: PostInputCategory;
  tags: string[];
  authorId: number;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  bike?: string | null;
}

export interface Comment {
  id: number;
  postId: number;
  authorName: string;
  body: string;
  createdAt: string;
}

export interface CommentInput {
  /**
     * @minLength 1
     * @maxLength 80
     */
  authorName: string;
  /**
     * @minLength 1
     * @maxLength 2000
     */
  body: string;
}

export interface TagCount {
  tag: string;
  count: number;
}

export type CategoryCountCategory = typeof CategoryCountCategory[keyof typeof CategoryCountCategory];


export const CategoryCountCategory = {
  'ride-report': 'ride-report',
  'gear-review': 'gear-review',
  'tech-tips': 'tech-tips',
  news: 'news',
} as const;

export interface CategoryCount {
  category: CategoryCountCategory;
  count: number;
}

export interface BlogStats {
  totalPosts: number;
  totalComments: number;
  totalLikes: number;
  totalAuthors: number;
  categories: CategoryCount[];
}

export type ListPostsParams = {
tag?: string;
category?: ListPostsCategory;
search?: string;
};

export type ListPostsCategory = typeof ListPostsCategory[keyof typeof ListPostsCategory];


export const ListPostsCategory = {
  'ride-report': 'ride-report',
  'gear-review': 'gear-review',
  'tech-tips': 'tech-tips',
  news: 'news',
} as const;

export type ListPopularPostsParams = {
/**
 * @minimum 1
 * @maximum 20
 */
limit?: number;
};