# lib/api-zod/src/generated/types/author.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BikerBlog API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Author {
  id: number;
  name: string;
  avatarUrl: string;
  bio: string;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  bike?: string | null;
}