# lib/api-zod/src/generated/types/categoryCount.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BikerBlog API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryCountCategory } from './categoryCountCategory';

export interface CategoryCount {
  category: CategoryCountCategory;
  count: number;
}