# lib/api-zod/src/generated/types/comment.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BikerBlog API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Comment {
  id: number;
  postId: number;
  authorName: string;
  body: string;
  createdAt: Date;
}