# lib/api-zod/src/generated/types/commentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BikerBlog API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CommentInput {
  /**
     * @minLength 1
     * @maxLength 80
     */
  authorName: string;
  /**
     * @minLength 1
     * @maxLength 2000
     */
  body: string;
}