# lib/api-zod/src/generated/types/post.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BikerBlog API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Author } from './author';
import type { PostCategory } from './postCategory';

export interface Post {
  id: number;
  slug: string;
  title: string;
  excerpt: string;
  /** Markdown body */
  content: string;
  coverImageUrl: string;
  category: PostCategory;
  tags: string[];
  author: Author;
  publishedAt: Date;
  readingMinutes: number;
  likeCount: number;
  commentCount: number;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  bike?: string | null;
}