# lib/api-zod/src/generated/types/postInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BikerBlog API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PostInputCategory } from './postInputCategory';

export interface PostInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  excerpt: string;
  /** @minLength 1 */
  content: string;
  coverImageUrl: string;
  category: PostInputCategory;
  tags: string[];
  authorId: number;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  bike?: string | null;
}